refactor(media): Add enum-identifiers.ts
